[core]: move occm package to pkg dir
